requirements.txt
tensorflow==2.12.0
pandas
dvc
mlflow==2.6.0
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
-e .